refactor(services): extract AnalyzeService from Orchestrator
Move both analyse modes -- analyze_bulk (single_pass) and
analyze_hierarchical -- together with their helpers out of the
Orchestrator god-class and into a new AnalyzeService in
qfa.services.analyze, per ADR-017 and epic #112.

Both modes stay on one service: they are two modes of one endpoint,
selected on the request, and they share the retained-placeholder
guardrail (_ANALYZE_RETAINED_PLACEHOLDER_TYPES), which stays a single
definition. AnalyzeService is also the only use case that needs an
EmbeddingPort, so that dependency now sits on its constructor instead
of on one shared by four use cases that never touch it.

The service has no base class and takes the shared LLMCallExecutor as
an injected collaborator. qfa.api.composition.build_services builds the
object graph once -- one executor and one PresidioAnonymizer shared by
every service -- with build_orchestrator and build_analyze_service as
thin single-service wrappers. The analyze route now annotates against
AnalyzeService via a new get_analyze_service provider.

Two helpers shared with the not-yet-extracted summarise path move to
neutral homes rather than being duplicated: hyperlink_form_references
to qfa.services.record_links and the judge user message to
qfa.services.prompts.

Analyse tests move to tests/services/test_analyze.py and
tests/services/test_analyze_hierarchical.py, constructing the real
LLMCallExecutor over FakeLLMPort / FakeAnonymizer. The e2e and API
tests for the analyze endpoint pass unmodified in both modes.

Refs #112
Closes #266
